backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
numpy
python-multipart
reportlab